WIP on refactor for pop on java broker
