+++ common/src/main/java/com/microsoft/identity/common/internal/platform/AndroidDeviceKeyManager.java
@@ -22,13 +22,18 @@
 // THE SOFTWARE.
 package com.microsoft.identity.common.internal.platform;
 
+import static com.microsoft.identity.common.java.exception.ClientException.INVALID_PROTECTION_PARAMS;
+import static com.microsoft.identity.common.java.exception.ClientException.KEYSTORE_NOT_INITIALIZED;
+import static com.microsoft.identity.common.java.exception.ClientException.NO_SUCH_ALGORITHM;
+import static com.microsoft.identity.common.java.exception.ClientException.UNKNOWN_ERROR;
+
 import android.os.Build;
 import android.security.keystore.KeyInfo;
 
-import com.microsoft.identity.common.java.crypto.IAndroidKeyStoreKeyManager;
 import com.microsoft.identity.common.java.crypto.SecureHardwareState;
 import com.microsoft.identity.common.java.exception.ClientException;
-import com.microsoft.identity.common.internal.util.Supplier;
+import com.microsoft.identity.common.java.platform.AbstractKeyStoreKeyManager;
+import com.microsoft.identity.common.java.util.Supplier;
 import com.microsoft.identity.common.logging.Logger;
 
 import java.security.KeyFactory;
@@ -37,23 +42,14 @@ import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
 import java.security.PrivateKey;
 import java.security.UnrecoverableEntryException;
-import java.security.cert.Certificate;
 import java.security.spec.InvalidKeySpecException;
-import java.util.Arrays;
-import java.util.Date;
 
 import javax.crypto.SecretKey;
 import javax.crypto.SecretKeyFactory;
 
 import lombok.Builder;
-import lombok.Getter;
 import lombok.NonNull;
 import lombok.experimental.Accessors;
-
-import static com.microsoft.identity.common.java.exception.ClientException.INVALID_PROTECTION_PARAMS;
-import static com.microsoft.identity.common.java.exception.ClientException.KEYSTORE_NOT_INITIALIZED;
-import static com.microsoft.identity.common.java.exception.ClientException.NO_SUCH_ALGORITHM;
-import static com.microsoft.identity.common.java.exception.ClientException.UNKNOWN_ERROR;
 
 /**
  * A manager class for providing access to a particular entry in a KeyStore.
@@ -61,129 +57,13 @@ import static com.microsoft.identity.common.java.exception.ClientException.UNKNO
  * @param <K> the type of KeyStore.Entry being managed.
  */
 @Accessors(prefix = "m")
-public class DeviceKeyManager<K extends KeyStore.Entry> implements IAndroidKeyStoreKeyManager<K> {
+public class AndroidDeviceKeyManager<K extends KeyStore.Entry> extends AbstractKeyStoreKeyManager<K> {
 
-    private static final String TAG = DeviceKeyManager.class.getSimpleName();
-    private final KeyStore mKeyStore;
-    @Getter
-    private final String mKeyAlias;
-    private final Supplier<byte[]> mThumbprintSupplier;
+    private static final String TAG = AndroidDeviceKeyManager.class.getSimpleName();
 
     @Builder
-    public DeviceKeyManager(@NonNull final KeyStore keyStore, @NonNull final String keyAlias,
-                            @NonNull final Supplier<byte[]> thumbprintSupplier) throws KeyStoreException {
-        this.mKeyAlias = keyAlias;
-        this.mThumbprintSupplier = thumbprintSupplier;
-        this.mKeyStore = keyStore;
-    }
-
-
-    @Override
-    public boolean exists() {
-        boolean exists = false;
-
-        try {
-            exists = mKeyStore.containsAlias(mKeyAlias);
-        } catch (final KeyStoreException e) {
-            Logger.error(
-                    TAG,
-                    "Error while querying KeyStore",
-                    e
-            );
-        }
-
-        return exists;
-
-    }
-
-    @Override
-    public boolean hasThumbprint(@NonNull final byte[] thumbprint) {
-        return Arrays.equals(thumbprint, mThumbprintSupplier.get());
-    }
-
-    @Override
-    public Date getCreationDate() throws ClientException {
-        try {
-            return mKeyStore.getCreationDate(mKeyAlias);
-        } catch (final KeyStoreException e) {
-            Logger.error(
-                    TAG,
-                    "Error while getting creation date for alias " + mKeyAlias,
-                    e
-            );
-            throw new ClientException(ClientException.KEYSTORE_NOT_INITIALIZED, e.getMessage(), e);
-        }
-    }
-
-    @Override
-    public boolean clear() {
-        boolean deleted = false;
-
-        try {
-            mKeyStore.deleteEntry(mKeyAlias);
-            deleted = true;
-        } catch (final KeyStoreException e) {
-            Logger.error(
-                    TAG,
-                    "Error while clearing KeyStore",
-                    e
-            );
-        }
-
-        return deleted;
-    }
-
-
-    /**
-     * Retrieve the entry stored in this particular alias.
-     *
-     * @return the Entry in question, or null if the entry does not exist.
-     * @throws UnrecoverableEntryException
-     * @throws NoSuchAlgorithmException
-     * @throws KeyStoreException
-     */
-    @SuppressWarnings("unchecked")
-    @Override
-    public K getEntry() throws UnrecoverableEntryException, NoSuchAlgorithmException, KeyStoreException {
-        return (K) mKeyStore.getEntry(mKeyAlias, null);
-    }
-
-    @Override
-    public void importKey(@NonNull final byte[] jwk, @NonNull final String algorithm) throws ClientException {
-        throw new UnsupportedOperationException("This is not currently supported");
-    }
-
-    @Override
-    public byte[] getThumbprint() {
-        return mThumbprintSupplier.get();
-    }
-
-    @Override
-    public Certificate[] getCertificateChain() throws ClientException {
-        final Exception exception;
-        final String errCode;
-
-        try {
-            return mKeyStore.getCertificateChain(mKeyAlias);
-        } catch (final KeyStoreException e) {
-            exception = e;
-            errCode = KEYSTORE_NOT_INITIALIZED;
-        }
-
-        final ClientException clientException = new ClientException(
-                errCode,
-                exception.getMessage(),
-                exception
-        );
-
-        Logger.error(
-                TAG,
-                clientException.getMessage(),
-                clientException
-        );
-
-        throw clientException;
-
+    public AndroidDeviceKeyManager(@NonNull final KeyStore keyStore, @NonNull final String keyAlias) throws KeyStoreException {
+        super(keyStore, keyAlias, null);
     }
 
     /**
