+++ common4j/src/main/com/microsoft/identity/common/java/util/Supplier.java
@@ -20,7 +20,9 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.common.internal.util;
+package com.microsoft.identity.common.java.util;
+
+import com.microsoft.identity.common.java.exception.ClientException;
 
 /**
  * Represents an operation that returns a result, taking no arguments.
